+++ routes/exam.route.js
@@ -11,14 +11,6 @@ router.post("/create", ensureAuthenticated, examController.createExam);
 router.get("/myExams", ensureAuthenticated, examController.getAllExams);
 router.get("/:id", examController.getExamById);
 router.post("/startExam", ensureAuthenticated, examController.startExam);
-// Q1: Should this route be in another route file like "question.route"?
-// Q2: ensureAuth?
-router.get(
-	"/:id/questions",
-	ensureAuthenticated,
-	examController.getExamQuestions
-);
-
 router.post(
 	"/finishExam",
 	ensureAuthenticated,
